Event_manager/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-bower==5.2.0
django-colorfield==0.4.1
django-crispy-forms==1.11.2
django-extensions==3.1.3
django-scheduler==0.9.3
icalendar==4.0.7
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
